Ultrasonic Speed adjuster
An Arduino based projected where we implemented an embedded system. Our
concept was based on the automotive field where we could artificially
create an automatic cruise control system which auto detects the speed
depending on how far you are from an obstruction. implemented  with a 3
pin ping, sensor arduino board, rbg led, and a piezo which all worked
dynamically visually ,audibility, and physically.